fix forks during rewind, allow bisect of failures
When rewinding, we sometimes encounter a legitimate fork at the height we roll back to.This was not well handled and could loop forever awaiting a height reduction that could not be reached.

Secondly, if a mismatch occurs, it is difficult to bisect it to a single or group of addresses. Allow the batch size to be overridden to allow a bisect operation, and verify that each chunk matches our expectations.